componentize DiscoverFeed
