+++ packages/apolloschurchapp/src/tabs/discover/DiscoverFeed/getContentChannels.js
